feat : Account Service ( 회원가입기능 ) 추가

+++ src/main/java/com/uos/all4runner/domain/dto/request/AccountRequest.java
@@ -17,7 +17,7 @@ public interface AccountRequest {
 		@NotBlank Gender gender,
 		@NotBlank String addressGu,
 		@NotBlank String addressDong,
-		@Min(value = 0) @Max(value = 20) Integer avgspeed,
+		@Min(value = 0) @Max(value = 20) Double avgspeed,
 		@Min(value = 0) @Max(value = 200) Integer weight
 		){
 	}
